[XDS Interop] Move XdsStatsWatcher to a separate file.
This will help with introducing test coverage as the logic becomes
more complex.
